requirements.txt
aiogram
APScheduler
python-decouple
redis
asyncpg-lite
apscheduler
markovify